PR comment #2/2 re-org files
